requirements.txt
# generated by rye
# use `rye lock` or `rye sync` to update this lockfile
#
# last locked with the following flags:
#   pre: false
#   features: []
#   all-features: false
#   with-sources: false
#   generate-hashes: false
#   universal: false

aiotraq==0.2.4
    # via aiotraq-message
aiotraq-bot==0.2.4
    # via aiotraq-message
    # via bot-kitchin-timer
aiotraq-message==0.2.4
    # via bot-kitchin-timer
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via httpx
    # via starlette
    # via watchfiles
attrs==23.2.0
    # via aiotraq
certifi==2024.7.4
    # via httpcore
    # via httpx
click==8.1.7
    # via typer
    # via uvicorn
dnspython==2.6.1
    # via email-validator
email-validator==2.2.0
    # via fastapi
fastapi==0.111.1
    # via aiotraq-bot
fastapi-cli==0.0.4
    # via fastapi
h11==0.14.0
    # via httpcore
    # via uvicorn
httpcore==1.0.5
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx==0.27.0
    # via aiotraq
    # via fastapi
idna==3.7
    # via anyio
    # via email-validator
    # via httpx
jinja2==3.1.4
    # via fastapi
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
numpy==1.26.4
    # via aiotraq-message
    # via pandas
pandas==2.2.2
    # via aiotraq-message
pillow==10.4.0
    # via aiotraq-message
pydantic==2.8.2
    # via fastapi
pydantic-core==2.20.1
    # via pydantic
pygments==2.18.0
    # via rich
python-dateutil==2.9.0.post0
    # via aiotraq
    # via pandas
python-dotenv==1.0.1
    # via bot-kitchin-timer
    # via uvicorn
python-multipart==0.0.9
    # via fastapi
pytz==2024.1
    # via pandas
pyyaml==6.0.1
    # via uvicorn
rich==13.7.1
    # via typer
shellingham==1.5.4
    # via typer
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
    # via httpx
starlette==0.37.2
    # via fastapi
tabulate==0.9.0
    # via aiotraq-message
typer==0.12.3
    # via fastapi-cli
    # via fastapi
    # via pydantic
    # via pydantic-core
    # via typer
tzdata==2024.1
    # via pandas
uvicorn==0.30.3
    # via aiotraq-bot
    # via fastapi
uvloop==0.19.0
    # via uvicorn
watchfiles==0.22.0
    # via uvicorn
websockets==12.0
    # via uvicorn
